+++ Views/Settings/AboutTab.axaml.cs
@@ -1,5 +1,7 @@
 using Avalonia.Controls;
+
 namespace NullWave.Views.Settings;
+
 public partial class AboutTab : UserControl
 {
     public AboutTab() => InitializeComponent();